gradio/requirements.txt
numpy==1.26.4
Pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
transformers==4.40.2
sentencepiece==0.1.99
https://github.com/Dao-AILab/flash-attention/releases/download/v2.6.2/flash_attn-2.6.2+cu123torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
gradio
decord
accelerate